patched errors

+++ iebank_api/routes.py
@@ -1,4 +1,4 @@
-from flask import Flask, request
+from flask import Flask, request, abort
 from iebank_api import db, app
 from iebank_api.models import Account
 
@@ -24,6 +24,10 @@ def skull():
 
 @app.route('/accounts', methods=['POST'])
 def create_account():
+    data = request.get_json()
+    required_fields = ['name', 'currency', 'country']
+    if not data or not all(field in data for field in required_fields):
+        abort(500)
     name = request.json['name']
     currency = request.json['currency']
     country = request.json['country']
@@ -42,11 +46,15 @@ def get_accounts():
 @app.route('/accounts/<int:id>', methods=['GET'])
 def get_account(id):
     account = Account.query.get(id)
+    if not account:
+        abort(500)
     return format_account(account)
 
 @app.route('/accounts/<int:id>', methods=['PUT'])
 def update_account(id):
     account = Account.query.get(id)
+    if not account:
+        abort(500)
     account.name = request.json['name']
     db.session.commit()
     return format_account(account)
@@ -54,6 +62,8 @@ def update_account(id):
 @app.route('/accounts/<int:id>', methods=['DELETE'])
 def delete_account(id):
     account = Account.query.get(id)
+    if not account:
+        abort(500)
     db.session.delete(account)
     db.session.commit()
     return format_account(account)
